requirements.txt
gradio>=3.0.0
torch>=2.0.0
torchvision
torchaudio
diffusers>=0.19.0
transformers>=4.30.0
accelerate
flask>=2.0.0
flask-cors>=3.0.0
Pillow>=9.0.0